fixed bugs with the imports for the models

+++ app/models/product.py
@@ -1,6 +1,6 @@
 from .db import db, environment, SCHEMA
-from user import User
-from app.models.category import Category
+from .user import User
+from .category import Category
 
 class Product(db.Model):
     __tablename__ = 'products'
@@ -10,4 +10,4 @@ class Product(db.Model):
     category_id = db.Column(db.Integer, db.ForeignKey('Category.id'), nullable=False)
     name = db.Column(db.String(50), nullable=False)
     description = db.Column(db.String(1000), nullable=False)
-    price = db.Column(db.Decimal, nullable=False)
+    price = db.Column(db.Float, nullable=False)
